requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6
pyjwt>=2.8.0
